requirements.txt
python-telegram-bot==20.7
pywinrm>=0.4.3
python-dotenv>=1.0.0